+++ src/Control/Carrier/Cut/Church.hs
@@ -34,6 +34,16 @@ import Data.Functor.Identity
 
 -- | Run a 'Cut' effect with continuations respectively interpreting 'pure' / '<|>', 'empty', and 'cutfail'.
 --
+-- @
+-- runCut cons nil fail ('pure' a '<|>' 'empty') = cons a nil
+-- @
+-- @
+-- runCut cons nil fail 'cutfail' = fail
+-- @
+-- @
+-- runCut cons nil fail ('call' 'cutfail') = nil
+-- @
+--
 -- @since 1.0.0.0
 runCut :: (a -> m b -> m b) -> m b -> m b -> CutC m a -> m b
 runCut cons nil fail (CutC runCutC) = runCutC cons nil fail
@@ -106,8 +116,8 @@ instance Algebra sig m => Algebra (Cut :+: NonDet :+: sig) (CutC m) where
     L (Call m)       -> runCut cons nil nil (hdl (m <$ ctx))
     R (L (L Empty))  -> nil
     R (L (R Choose)) -> cons (True <$ ctx) (cons (False <$ ctx) nil)
-    R (R other)      -> thread (dst ~<~ hdl) other (pure ctx) >>= runIdentity . runCut (coerce cons) (coerce nil) (coerce fail)
+    R (R other)      -> thread (dst ~<~ hdl) other (pure ctx) >>= run . runCut (coerce cons) (coerce nil) (coerce fail)
     where
     dst :: Applicative m => CutC Identity (CutC m a) -> m (CutC Identity a)
-    dst = runIdentity . runCut (fmap . liftA2 (<|>) . runCut (fmap . (<|>) . pure) (pure empty) (pure cutfail)) (pure (pure empty)) (pure (pure cutfail))
+    dst = run . runCut (fmap . liftA2 (<|>) . runCut (fmap . (<|>) . pure) (pure empty) (pure cutfail)) (pure (pure empty)) (pure (pure cutfail))
   {-# INLINE alg #-}