+++ src/Control/Carrier/State/Church.hs
@@ -23,63 +23,74 @@ module Control.Carrier.State.Church
 ) where
 
 import Control.Algebra
-import Control.Applicative (Alternative(..))
+import Control.Applicative (Alternative(..), liftA2)
 import Control.Effect.State
-import Control.Monad (MonadPlus(..))
+import Control.Monad (MonadPlus)
 import Control.Monad.Fail as Fail
 import Control.Monad.Fix
 import Control.Monad.IO.Class
 import Control.Monad.Trans.Class
 
--- | Run a 'State' effect starting from the passed value.
+-- | Run a 'State' effect starting from the passed value, applying a continuation to the final state and result.
 --
 -- @
--- 'runState' s ('pure' a) = 'pure' (s, a)
+-- 'runState' k s ('pure' a) = k s a
 -- @
 -- @
--- 'runState' s 'get' = 'pure' (s, s)
+-- 'runState' k s 'get' = k s s
 -- @
 -- @
--- 'runState' s ('put' t) = 'pure' (t, ())
+-- 'runState' k s ('put' t) = k t ()
 -- @
 --
 -- @since 1.1.0.0
-runState :: forall s m a . Applicative m => s -> StateC s m a -> m (s, a)
-runState s (StateC m) = m (\ a s -> pure (s, a)) s
+runState :: forall s m a b . (s -> a -> m b) -> s -> StateC s m a -> m b
+runState f s (StateC m) = m f s
 {-# INLINE runState #-}
 
 -- | Run a 'State' effect, yielding the result value and discarding the final state.
 --
 -- @
--- 'evalState' s m = 'fmap' 'snd' ('runState' s m)
+-- 'evalState' = 'runState' ('const' 'pure')
 -- @
 --
 -- @since 1.1.0.0
 evalState :: forall s m a . Applicative m => s -> StateC s m a -> m a
-evalState s (StateC m) = m (const . pure) s
+evalState = runState (const pure)
 {-# INLINE evalState #-}
 
 -- | Run a 'State' effect, yielding the final state and discarding the return value.
 --
 -- @
--- 'execState' s m = 'fmap' 'fst' ('runState' s m)
+-- 'execState' = 'runState' ('const' '.' 'pure')
 -- @
 --
 -- @since 1.1.0.0
 execState :: forall s m a . Applicative m => s -> StateC s m a -> m s
-execState s (StateC m) = m (const pure) s
+execState = runState (const . pure)
 {-# INLINE execState #-}
 
 -- | @since 1.1.0.0
-newtype StateC s m a = StateC { runStateC :: forall r . (a -> s -> m r) -> s -> m r }
+newtype StateC s m a = StateC (forall r . (s -> a -> m r) -> s -> m r)
   deriving (Functor)
 
 instance Applicative (StateC s m) where
-  pure a = StateC $ \ k s -> k a s
+  pure a = StateC $ \ k s -> k s a
   {-# INLINE pure #-}
 
-  StateC f <*> StateC a = StateC $ \ k -> f (a . (k .))
+  StateC f <*> StateC a = StateC $ \ k -> f (\ s f' -> a (\ s' -> k s' . f') s)
   {-# INLINE (<*>) #-}
+
+  liftA2 f (StateC a) (StateC b) = StateC $ \ k ->
+    a (\ s' a' -> b (\ s'' -> k s'' . f a') s')
+  {-# INLINE liftA2 #-}
+
+  StateC a *> StateC b = StateC $ \ k -> a (const . b k)
+  {-# INLINE (*>) #-}
+
+  StateC a <* StateC b = StateC $ \ k ->
+    a (\ s' a' -> b (\ s'' _ -> k s'' a') s')
+  {-# INLINE (<*) #-}
 
 instance Alternative m => Alternative (StateC s m) where
   empty = StateC $ \ _ _ -> empty
@@ -89,7 +100,7 @@ instance Alternative m => Alternative (StateC s m) where
   {-# INLINE (<|>) #-}
 
 instance Monad (StateC s m) where
-  StateC a >>= f = StateC $ \ k -> a (\ a' -> runStateC (f a') k)
+  StateC a >>= f = StateC $ \ k -> a (\ s -> runState k s . f)
   {-# INLINE (>>=) #-}
 
 instance Fail.MonadFail m => Fail.MonadFail (StateC s m) where
@@ -97,7 +108,7 @@ instance Fail.MonadFail m => Fail.MonadFail (StateC s m) where
   {-# INLINE fail #-}
 
 instance MonadFix m => MonadFix (StateC s m) where
-  mfix f = StateC $ \ k s -> mfix (runState s . f . snd) >>= uncurry (flip k)
+  mfix f = StateC $ \ k s -> mfix (runState (curry pure) s . f . snd) >>= uncurry k
   {-# INLINE mfix #-}
 
 instance MonadIO m => MonadIO (StateC s m) where
@@ -107,12 +118,12 @@ instance MonadIO m => MonadIO (StateC s m) where
 instance (Alternative m, Monad m) => MonadPlus (StateC s m)
 
 instance MonadTrans (StateC s) where
-  lift m = StateC $ \ k s -> m >>= flip k s
+  lift m = StateC $ \ k s -> m >>= k s
   {-# INLINE lift #-}
 
 instance Algebra sig m => Algebra (State s :+: sig) (StateC s m) where
   alg hdl sig ctx = StateC $ \ k s -> case sig of
-    L Get     -> k (s <$ ctx) s
-    L (Put s) -> k       ctx  s
-    R other   -> thread (uncurry runState ~<~ hdl) other (s, ctx) >>= uncurry (flip k)
+    L Get     -> k s (s <$ ctx)
+    L (Put s) -> k s       ctx
+    R other   -> thread (uncurry (runState (curry pure)) ~<~ hdl) other (s, ctx) >>= uncurry k
   {-# INLINE alg #-}