+++ test/Empty.hs
@@ -10,8 +10,9 @@ module Empty
 , test
 ) where
 
-import qualified Control.Carrier.Empty.Maybe as EmptyC
-import qualified Control.Monad.Trans.Maybe as MaybeT
+import qualified Control.Carrier.Empty.Church as C.Church
+import qualified Control.Carrier.Empty.Maybe as C.Maybe
+import qualified Control.Monad.Trans.Maybe as T.Maybe
 import           Control.Effect.Empty
 import           Data.Maybe (maybeToList)
 import           Gen
@@ -22,12 +23,17 @@ import           Test.Tasty.Hedgehog
 
 tests :: TestTree
 tests = testGroup "Empty"
-  [ testGroup "EmptyC" $
+  [ testGroup "EmptyC (Church)" $
     [ testMonad
     , testMonadFix
     , testEmpty
-    ] >>= ($ runL (fmap maybeToList . EmptyC.runEmpty))
-  , testGroup "MaybeT" $ testEmpty (runL (fmap maybeToList . MaybeT.runMaybeT))
+    ] >>= ($ runL (fmap maybeToList . C.Church.runEmpty (pure Nothing) (pure . Just)))
+  , testGroup "EmptyC (Maybe)" $
+    [ testMonad
+    , testMonadFix
+    , testEmpty
+    ] >>= ($ runL (fmap maybeToList . C.Maybe.runEmpty))
+  , testGroup "MaybeT" $ testEmpty (runL (fmap maybeToList . T.Maybe.runMaybeT))
   , testGroup "Maybe"  $ testEmpty (runL (pure . maybeToList))
   ] where
   testMonad    run = Monad.test    (m gen0 (\ _ _ -> [])) a b c initial run